removed nodemon which errored
