+++ package.json
@@ -14,7 +14,7 @@
   "main": "dist/index.js",
   "scripts": {
     "babel-node": "babel-node",
-    "start": "nodemon --exec yarn babel-node -- ./example/example.js",
+    "start": "pm2 start ./example/example.js --watch",
     "test": "jest --notify",
     "build-server": "babel ./index.js --out-dir ./dist && babel ./lib --out-dir ./dist/lib && babel ./example --out-dir ./dist/example",
     "build": "yarn clean && yarn build-server",
@@ -37,6 +37,7 @@
     "@babel/node": "^7.2.2",
     "@babel/plugin-proposal-object-rest-spread": "^7.4.3",
     "@babel/preset-env": "^7.4.3",
+    "@babel/register": "^7.4.0",
     "babel-eslint": "^10.0.1",
     "babel-jest": "^24.7.1",
     "eslint": "^5.16.0",
@@ -44,7 +45,7 @@
     "eslint-config-prettier": "^4.1.0",
     "eslint-plugin-import": "^2.17.2",
     "jest": "^24.7.1",
-    "nodemon": "^1.18.11",
+    "pm2": "^3.5.0",
     "prettier": "^1.17.0",
     "rimraf": "^2.6.3"
   },
